Add G_PARAM_NO_VALIDATION
This flag will be used to indicate that
g_param_value_validate will never change
a value for this pspec. That is the case
for all string properties, and it can be
safely set for boolean properties too. The
next commit will use this to speed up
setting of properties.